preparing for merge with main, disconnect working so far
