+++ examples/Manifest.toml
@@ -1,8 +1,8 @@
 # This file is machine-generated - editing it directly is not advised
 
-julia_version = "1.11.2"
+julia_version = "1.11.1"
 manifest_format = "2.0"
-project_hash = "caad15773f847e89193ba2659df97e05be9c0262"
+project_hash = "d9e6149740aa8d3e9a8390b693a97f5c7ff1bf11"
 
 [[deps.AbstractFFTs]]
 deps = ["LinearAlgebra"]
@@ -371,7 +371,7 @@ version = "1.2.1"
 
 [[deps.FerriteGridUtil]]
 deps = ["Ferrite", "HDF5", "Makie", "OrderedCollections"]
-path = "D:\\FerriteGridUtil.jl"
+path = "C:\\IAM\\FerriteGridUtil.jl"
 uuid = "bab52d2a-b59e-463f-9025-b61488042faa"
 version = "1.0.0-DEV"
 
@@ -987,6 +987,12 @@ version = "2.28.6+0"
 deps = ["ColorTypes", "FileIO", "GeometryBasics", "Printf"]
 git-tree-sha1 = "14a12d9153b1a1a22d669eede58b2ea2164ff138"
 uuid = "7269a6da-0436-5bbc-96c2-40638cbb6118"
+version = "0.4.13"
+
+[[deps.MethodAnalysis]]
+deps = ["AbstractTrees"]
+git-tree-sha1 = "c2ee9b8f036c951f9ed0a47503a7f7dc0905b256"
+uuid = "85b6ec6f-f7df-4429-9514-a64bcd9ee824"
 version = "0.4.13"
 
 [[deps.MicrosoftMPI_jll]]
