+++ examples/Project.toml
@@ -5,6 +5,7 @@ Ferrite = "c061ca5d-56c9-439f-9c0e-210fe06d3992"
 FerriteGmsh = "4f95f4f8-b27c-4ae5-9a39-ea55e634e36b"
 FerriteGridUtil = "bab52d2a-b59e-463f-9025-b61488042faa"
 GLMakie = "e9467ef8-e4e7-5192-8a1a-b1aee30e663a"
+MethodAnalysis = "85b6ec6f-f7df-4429-9514-a64bcd9ee824"
 OrderedCollections = "bac558e1-5e72-5ebc-8fee-abe8a469f55d"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 Tensors = "48a634ad-e948-5137-8d70-aa71f2a747f4"
